fix(ui): wait for restored draft before clearing
